fix some errors

+++ engine/src/engine/runtime/render_system.cpp
@@ -53,7 +53,7 @@ static KernelData GenerateSsaoKernel() {
 
     const int kernel_size = 32;
     const float inv_kernel_size = 1.0f / kernel_size;
-    for (int i = 0; i < kernel.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(kernel.size()); ++i) {
         // [-1, 1], [-1, 1], [0, 1]
         Vector3f sample(Random::Float(-1.0f, 1.0f),
                         Random::Float(-1.0f, 1.0f),
